ion: Count contiguous regions as large pages in debugfs
The page breakdown of ION buffers was added to debugfs in d60c07983d29.
However, due to the way buffers are allocated and the results observed
we became suspicious that many low order pages were in fact physically
contiguous and could be counted as a higher order for the purpose of our
on-going analysis.

This commit changes the way buffers are accounted so contiguous regions
in a buffer go into the highest page order possible even if the sg_list
is made up of low order entries (Provided alignment at the higher page
order is met).

Bug: 120440504

Test: cat /d/ion/heaps/* page breakdown looks more evenly distributed

Sample:

-----------------------       4K       8K      16K      32K      64K     128K     256K     512K       1M       2M       4M      >=8M

Before
ndroid.systemui             5979        0        0        0     1277        0        0        0        0        1        0        0M

After:
ndroid.systemui               34       21       12        5        7        9        7        5        3        2       19        0M

Change-Id: I7136979e8d8cca130c512633c60e64ec3804cd63
Signed-off-by: Erick Reyes <erickreyes@google.com>
Signed-off-by: Sorayukii <sorayukii69@gmail.com>