sched: features: Allow aggressive remote task interruptions
Add a sched feature (TTWU_QUEUE_SAME_FORCE) that will allow remote task
interrupts to be queued on the origin CPU even if the current CPU and
origin CPU share a cache. And also do not wake up and idle CPU for the chance of a cache hit. This avoids
unnecessary energy consumption. Plus, if we need to wake an idle CPU to
handle a remote wakeup, we will induce additional scheduling latency
since the remote CPU will be unavailable while leaving an idle state.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Sorayukii <sorayukii69@gmail.com>